gremlin_gplv3/conftest.py
import pytest

from gremlin_gplv3.users.models import User
from gremlin_gplv3.users.tests.factories import UserFactory


@pytest.fixture(autouse=True)
def media_storage(settings, tmpdir):
    settings.MEDIA_ROOT = tmpdir.strpath


@pytest.fixture
def user() -> User:
    return UserFactory()
