requirements/base.txt
pytz==2020.1  # https://github.com/stub42/pytz
python-slugify==4.0.0  # https://github.com/un33k/python-slugify
Pillow==7.1.2  # https://github.com/python-pillow/Pillow
argon2-cffi==19.2.0  # https://github.com/hynek/argon2_cffi
whitenoise==5.0.1  # https://github.com/evansd/whitenoise
redis==3.5.0  # https://github.com/andymccurdy/redis-py
hiredis==1.0.1  # https://github.com/redis/hiredis-py
celery==4.4.2  # pyup: < 5.0  # https://github.com/celery/celery
django-celery-beat==2.0.0  # https://github.com/celery/django-celery-beat
flower==0.9.4  # https://github.com/mher/flower
boto3==1.10.44
botocore==1.13.44

# Django
# ------------------------------------------------------------------------------
django==3.0.5  # pyup: < 3.1  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==4.0.0  # https://github.com/jazzband/django-model-utils
django-allauth==0.41.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.9.0  # https://github.com/django-crispy-forms/django-crispy-forms
django-redis==4.11.0  # https://github.com/jazzband/django-redis
django-cors-headers==3.2.0
django-cleanup==4.0.0
django-filter==2.2.0
django-url-filter==0.3.14
drf-yasg==1.17.1 #https://github.com/axnsan12/drf-yasg

# Django REST Framework
djangorestframework==3.11.0  # https://github.com/encode/django-rest-framework
djangorestframework-simplejwt==4.4.0
djangorestframework-api-key==2.0.0
djangorestframework-bulk==0.2.1
django-storages==1.8
PyJWT==1.7.1

# Data Science and Doc Extract
tika>=1.22
Cython>=0.29.19
scikit-learn>=0.21.3
numpy>=1.17.0
pandas>=0.25.0
jsonschema>=3.2.0
docutils>=0.15.2
docx2txt>=0.8
